feat(inngest): execute objectives durably with retry-backed agent runs, idempotent persistence and cap tool results echoed to the model while persisting full outputs

+++ packages/workflow/src/schemas.ts
@@ -44,7 +44,7 @@ export const toolCallRecordSchema = z.object({
 			id: z.string().min(1),
 			operation: z.string().optional(),
 			requestId: z.string().optional(),
-			durationMs: z.number().int().nonnegative().optional(),
+			durationMs: z.number().nonnegative().optional(),
 		})
 		.optional(),
 	startedAt: z.coerce.date(),
